+++ test/cases/test_misc.py
@@ -74,17 +74,17 @@ def test_example():
 
 # Debug
 
-def test_midi_messages():
-    midi_file = MidiFile.open_midi_file(Path(__file__).parent.parent.joinpath("res").joinpath("subject.mid"))
-
-    for i, track in enumerate(midi_file.tracks):
-        print(f"Track {i + 1}")
-        for j, message in enumerate(track.messages):
-            print(message)
-
-
-def test_subject():
-    sequences = Sequence.from_midi_file(Path(__file__).parent.parent.joinpath("res").joinpath("subject.mid"))
-    sequence = sequences[0]
-    sequence.merge(sequences[1:])
-    sequence.save("subject_out.mid")
+# def test_midi_messages():
+#     midi_file = MidiFile.open_midi_file(Path(__file__).parent.parent.joinpath("res").joinpath("subject.mid"))
+#
+#     for i, track in enumerate(midi_file.tracks):
+#         print(f"Track {i + 1}")
+#         for j, message in enumerate(track.messages):
+#             print(message)
+#
+#
+# def test_subject():
+#     sequences = Sequence.from_midi_file(Path(__file__).parent.parent.joinpath("res").joinpath("subject.mid"))
+#     sequence = sequences[0]
+#     sequence.merge(sequences[1:])
+#     sequence.save("subject_out.mid")